Add blogs controller and enhance validation on blog service

+++ BlogSite.Infrastructure/Common/Foundations/BlogService.cs
@@ -3,6 +3,7 @@ using BlogSite.Domain.Entities;
 using BlogSite.Persistence.Repositories.Interfaces;
 using System.ComponentModel.DataAnnotations;
 using System.Linq.Expressions;
+using System.Security.Authentication;
 
 namespace BlogSite.Infrastructure.Common.Foundations;
 
@@ -39,18 +40,37 @@ public class BlogService : IBlogService
         var foundBlog = await GetByIdAsync(blog.Id, cancellationToken: cancellationToken)
             ?? throw new InvalidOperationException("Blog not found!");
 
+        if (foundBlog.AuthorId != blog.AuthorId)
+            throw new AuthenticationException("Forbidden action!");
+
         foundBlog.Title = blog.Title;
         foundBlog.Content = blog.Content;
         foundBlog.ModifiedDate = DateTimeOffset.UtcNow;
 
-        return await _blogRepository.CreateAsync(blog, saveChanges, cancellationToken);
+        return await _blogRepository.UpdateAsync(foundBlog, saveChanges, cancellationToken);
     }
 
-    public async ValueTask<Blog> DeleteAsync(Blog blog, bool saveChanges = true, CancellationToken cancellationToken = default) =>
-        await _blogRepository.DeleteAsync(blog, saveChanges, cancellationToken);
+    public async ValueTask<Blog> DeleteAsync(Blog blog, bool saveChanges = true, CancellationToken cancellationToken = default)
+    {
+        var foundBlog = await GetByIdAsync(blog.Id, true, cancellationToken)
+            ?? throw new InvalidOperationException("Blog not found!");
 
-    public async ValueTask<Blog> DeleteByIdAsync(Guid id, bool saveChanges = true, CancellationToken cancellationToken = default) =>
-        await _blogRepository.DeleteByIdAsync(id, saveChanges, cancellationToken);
+        if (foundBlog.AuthorId != blog.AuthorId)
+            throw new AuthenticationException("Forbidden action!");
+
+        return await _blogRepository.DeleteAsync(blog, saveChanges, cancellationToken);
+    }
+
+    public async ValueTask<Blog> DeleteByIdAsync(Guid id, Guid authorId, bool saveChanges = true, CancellationToken cancellationToken = default)
+    {
+        var foundBlog = await GetByIdAsync(id, true, cancellationToken)
+            ?? throw new InvalidOperationException("Blog not found!");
+
+        if (foundBlog.AuthorId != authorId)
+            throw new AuthenticationException("Forbidden action!");
+
+        return await _blogRepository.DeleteByIdAsync(id, saveChanges, cancellationToken);
+    }
 
     private static bool IsValidBlog(Blog blog)
         => !(string.IsNullOrWhiteSpace(blog.Title)
